Fixed failing test in logging package
The configuration package was changed which meant in such a
way that the config struct was initialised with all 0 values. This
included a 0 value for the info and error log files.

I just added default values to the error and info log sections so that
the tests open a dummy log file.