Refactored the structure of the commands package
This refactor defines functions which allow the user to define their
command without having to worry about the discord side of things.

This makes the individual command functions unit-testable.

Users also no longer have to worry about whether or not permission
exists for a user to run a command. Note that if the command is not
mentioned in the config, it will assume that permission exusts.

Command functions are moved to `command_funcs.go`. This file will hold
all new commands and the corresponding `_test.go` fill will be the command
unit tests.